+++ modules/export/sql.js
@@ -219,13 +219,18 @@ const formatValue = (meta, config, currentValue, valueSrc, valueType, fieldWidge
       }
       ret = fn(...args);
     } else {
-      ret = SqlString.escape(currentValue);
+      if (Array.isArray(currentValue)) {
+        ret = currentValue.map(v => SqlString.escape(v));
+      } else {
+        ret = SqlString.escape(currentValue);
+      }
     }
   }
   return ret;
 };
 
 const formatField = (meta, config, field) => {
+  if (!field) return;
   const {fieldSeparator} = config.settings;
   const fieldDefinition = getFieldConfig(config, field) || {};
   const fieldParts = Array.isArray(field) ? field : field.split(fieldSeparator);