+++ modules/import/spel.js
@@ -36,7 +36,7 @@ export const _loadFromSpel = (spelStr, config, returnErrors = true) => {
 
     jsTree = convertToTree(convertedObj, conv, extendedConfig, meta);
     if (jsTree && jsTree.type != "group" && jsTree.type != "switch_group") {
-      jsTree = wrapInDefaultConj(jsTree, extendedConfig);
+      jsTree = wrapInDefaultConj(jsTree, extendedConfig, convertedObj["not"]);
     }
     logger.debug("jsTree:", jsTree);
   }
@@ -593,8 +593,10 @@ const convertToTree = (spel, conv, config, meta, parentSpel = null) => {
       opKey = "is_not_null";
     } else if (op == "le" && spel.children[1].type == "string" && spel.children[1].val == "") {
       opKey = "is_empty";
+      opKeys = ["is_empty"];
     } else if (op == "gt" && spel.children[1].type == "string" && spel.children[1].val == "") {
       opKey = "is_not_empty";
+      opKeys = ["is_not_empty"];
     } else if (op == "between") {
       opKey = "between";
       opKeys = ["between"];
@@ -673,8 +675,9 @@ const convertToTree = (spel, conv, config, meta, parentSpel = null) => {
 
           //todo: it's naive
           const widgets = opKeys.map(op => ({op, widget: getWidgetForFieldOp(config, field, op)}));
-          if (op == "eq") {
-            const ws = widgets.find(({op, widget}) => (widget != "field"));
+          
+          if (op == "eq" || op == "ne") {
+            const ws = widgets.find(({ op, widget }) => (widget && widget != "field"));
             opKey = ws.op;
           }
         }
@@ -861,7 +864,7 @@ const wrapInDefaultConj = (rule, config, not = false) => {
     children1: { [rule.id]: rule },
     properties: {
       conjunction: defaultConjunction(config),
-      not: not
+      not: not || false
     }
   };
 };